refactor(capture): extract JSONL terminal proofs
